pyproject.toml
[project]
name = "memo-bot-pro"
version = "1.0.0"
description = "Binance Advisory & Live Trading assistant with Telegram integration"
authors = [{name = "MeMo", email = "support@memobotpro.com"}]
license = {text = "Proprietary"}
requires-python = ">=3.11"
dependencies = [
    "python-binance==1.0.19",
    "requests==2.32.3",
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0.0"
]

[project.urls]
Homepage = "https://yourwebsite.com"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
